requirements.txt
Adafruit-Blinka==6.3.2
adafruit-circuitpython-ads1x15==2.2.8
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-motor==3.2.8
adafruit-circuitpython-motorkit==1.6.2
adafruit-circuitpython-pca9685==3.3.5
adafruit-circuitpython-register==1.9.5
Adafruit-PlatformDetect==3.2.0
Adafruit-PureIO==1.1.8
appdirs==1.4.4
attrs==20.3.0
black==20.8b1
click==7.1.2
flake8==3.8.4
importlib-metadata==3.7.0
iniconfig==1.1.1
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyftdi==0.52.9
pyparsing==2.4.7
pyserial==3.5
pytest==6.2.2
pyusb==1.1.1
regex==2020.11.13
smbus==1.1.post2
rpi-ws281x==4.2.5
RPi.GPIO==0.7.0
sysv-ipc==1.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
zipp==3.4.1
